Fix(viewChangeEnd): record new bounds for newly added markers
otherwise `_currentShownBounds` keeps the value that was last recorded when switching the "moveend" and "zoomend" event listeners (i.e. when the MCG is getting frozen), and dynamically added markers are displayed (or hidden) based on that possibly obsolete value, whereas the current map viewport may have changed (due to user zooming and/or panning the map).

Fix #7 (Markers disapear when adding them to a freezable group)