Refactor(example): use manage-libs-versions, urijs and load-js-css
so that it is much easier to demo and test with different Leaflet and Leaflet.markercluster versions, as well as MCG.Freezable versions.
Also separated some setup / initialization and main page code into independent script files, for clarity.
Use HTTPS for OSM tiles as well, even though the server looks to be currently unavailable…